+++ Sources/Controllers/TargetMenu/GPX/TrackMenu/OATrackMenuHudViewControllerConstants.h
@@ -1,0 +1,22 @@
+//
+//  OATrackMenuHudViewControllerConstants.h
+//  OsmAnd Maps
+//
+//  Created by nnngrach on 18.04.2023.
+//  Copyright © 2023 OsmAnd. All rights reserved.
+//
+
+typedef NS_ENUM(NSUInteger, EOATrackMenuHudTab)
+{
+    EOATrackMenuHudOverviewTab = 0,
+    EOATrackMenuHudSegmentsTab,
+    EOATrackMenuHudPointsTab,
+    EOATrackMenuHudActionsTab
+};
+
+typedef NS_ENUM(NSUInteger, EOATrackMenuHudSegmentsStatisticsTab)
+{
+    EOATrackMenuHudSegmentsStatisticsOverviewTab = 0,
+    EOATrackMenuHudSegmentsStatisticsAlititudeTab,
+    EOATrackMenuHudSegmentsStatisticsSpeedTab
+};
